mc: music volume is signed, the 0x5C0 flag unsigned
The options screen reads musicVolume signed like its neighbouring settings,
and both readers of unknown_5C0 load it unsigned; with those types the
options and bridge code reach the save's fields as members.

+++ src/main/loadui.c
@@ -235,9 +235,9 @@ void Snd_ApplyVolumeTable(s32 arg0)
             SndEvt_EnqueueType5(0, (u8)D_8007A396);
         }
     } else {
-        temp       = sp10.data[Mc_SaveData.musicVolume];
+        temp       = sp10.data[(u8)Mc_SaveData.musicVolume];
         D_8007A396 = temp;
-        if ((s8)Mc_SaveData.musicVolume == 3) {
+        if (Mc_SaveData.musicVolume == 3) {
             SndEvt_EnqueueType5Pending();
         } else {
             SndEvt_FlushType5Pending();
@@ -245,7 +245,7 @@ void Snd_ApplyVolumeTable(s32 arg0)
         if (gStageRoomSong != 0) {
             SndEvt_EnqueueType5(gStageRoomSong, (u8)D_8007A396);
         } else {
-            SndEvt_EnqueueType5(0, sp10.data[Mc_SaveData.musicVolume]);
+            SndEvt_EnqueueType5(0, sp10.data[(u8)Mc_SaveData.musicVolume]);
         }
     }
 }
